Public Functions ut 22

+++ ops/tests/ut/aicpu_test/testcase/get_dynamic_dims_utest.cpp
@@ -28,12 +28,12 @@ TEST_F(GET_DYNAMIC_DIMS_KERNEL_UT, INT32_Success)
       4, 3, 2, -1, 1, 3, 1, 2, 1, 5, 16, -1, -1, 3, 4 };
 
   NodeDefBuilder(node_def.get(), "GetDynamicDims", "GetDynamicDims")
-    .Input({"x1", DT_INT32, {4}, x1.data()})
-    .Input({"x2", DT_INT32, {3}, x2.data()})
-    .Input({"x3", DT_INT32, {5}, x3.data()})
-    .Output({"dims", DT_INT64, {3}, dims.data()})
-    .Attr("N", 3)
-    .Attr("shape_info", shape_info_attr);
+      .Input({"x1", DT_INT32, {4}, x1.data()})
+      .Input({"x2", DT_INT32, {3}, x2.data()})
+      .Input({"x3", DT_INT32, {5}, x3.data()})
+      .Output({"dims", DT_INT64, {3}, dims.data()})
+      .Attr("N", 3)
+      .Attr("shape_info", shape_info_attr);
 
   RUN_KERNEL(node_def, HOST, KERNEL_STATUS_OK);
 
@@ -53,12 +53,12 @@ TEST_F(GET_DYNAMIC_DIMS_KERNEL_UT, INT64_Success)
       4, 3, 2, -1, 1, 3, 1, 2, 1, 5, 16, -1, -1, 3, 4 };
 
   NodeDefBuilder(node_def.get(), "GetDynamicDims", "GetDynamicDims")
-    .Input({"x1", DT_INT32, {4}, x1.data()})
-    .Input({"x2", DT_INT32, {3}, x2.data()})
-    .Input({"x3", DT_INT32, {5}, x3.data()})
-    .Output({"dims", DT_INT64, {3}, dims.data()})
-    .Attr("N", 3)
-    .Attr("shape_info", shape_info_attr);
+      .Input({"x1", DT_INT64, {4}, x1.data()})
+      .Input({"x2", DT_INT64, {3}, x2.data()})
+      .Input({"x3", DT_INT64, {5}, x3.data()})
+      .Output({"dims", DT_INT64, {3}, dims.data()})
+      .Attr("N", 3)
+      .Attr("shape_info", shape_info_attr);
 
   RUN_KERNEL(node_def, HOST, KERNEL_STATUS_OK);
 
@@ -78,13 +78,13 @@ TEST_F(GET_DYNAMIC_DIMS_KERNEL_UT, OutputNum_Fail)
       4, 3, 2, -1, 1, 3, 1, 2, 1, 5, 16, -1, -1, 3, 4 };
 
   NodeDefBuilder(node_def.get(), "GetDynamicDims", "GetDynamicDims")
-    .Input({"x1", DT_INT32, {4}, x1.data()})
-    .Input({"x2", DT_INT32, {3}, x2.data()})
-    .Input({"x3", DT_INT32, {5}, x3.data()})
-    .Output({"dims", DT_INT64, {3}, dims.data()})
-    .Output({"dims2", DT_INT64, {3}, dims.data()})
-    .Attr("N", 3)
-    .Attr("shape_info", shape_info_attr);
+      .Input({"x1", DT_INT32, {4}, x1.data()})
+      .Input({"x2", DT_INT32, {3}, x2.data()})
+      .Input({"x3", DT_INT32, {5}, x3.data()})
+      .Output({"dims", DT_INT64, {3}, dims.data()})
+      .Output({"dims2", DT_INT64, {3}, dims.data()})
+      .Attr("N", 3)
+      .Attr("shape_info", shape_info_attr);
 
   RUN_KERNEL(node_def, HOST, KERNEL_STATUS_PARAM_INVALID);
 }
